scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Upwork vacancy scraper — headless Chrome + nodriver."
requires-python = ">=3.13"
dependencies = [
    "nodriver>=0.28",
    "psycopg2-binary>=2.9",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
